All working (i hope), note:ches variant, no en passant or 3 repetition rule

+++ Chess/src/businessLogic/ManagePlayerTurn.java
@@ -3,18 +3,18 @@ import data.Piece;
 import java.util.ArrayList;
 
 public class ManagePlayerTurn {
-    private static byte turn=0;
+    private static int turn=0;
     
     public static void changeTurn(){
         if(turn==1)turn=0;//switch turn
         else turn=1;
     }
 
-    public static byte getTurn() {
+    public static int getTurn() {
         return turn;
     }
 
-    public static void setTurn(byte aTurn) {
+    public static void setTurn(int aTurn) {
         turn = aTurn;
     }
 }
